% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_themes_as_list.r
\name{export_themes_as_list}
\alias{export_themes_as_list}
\title{Export RStudio Theme List as an Unstructed HTML list}
\usage{
export_themes_as_list(path = "assets/list_for_website.html")
}
\arguments{
\item{path}{specify the output path where the HTML file should be saved}
}
\value{
none
}
\description{
Exports the RStudio theme list as a HTML list that can be embedded into
websites
}
\examples{
\dontrun{
export_themes_as_list(path = "output.html")
}
}
