% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_themes_for_websites.r
\name{export_themes_for_website}
\alias{export_themes_for_website}
\title{Export RStudio Theme List as HTML table}
\usage{
export_themes_for_website(path = "assets/table_for_website.html")
}
\arguments{
\item{path}{specify the output path where the HTML file should be saved}
}
\value{
none
}
\description{
Exports the RStudio theme list as a HTML table that can be embedded into
websites, using the flextable package
}
\examples{
\dontrun{
export_themes_for_website(path = "output.html")
}
}
