% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_rstudio_themes.R
\name{install_rstudio_themes}
\alias{install_rstudio_themes}
\title{Add a Custom RStudio Editor Theme}
\usage{
install_rstudio_themes(theme, globally = FALSE)
}
\arguments{
\item{theme}{name of the theme to install, can be a character vector with multiple themes. the shortcode "all" will install all themes in the dataset; similarly, "all_light" and "all_dark" will install all light or dark themes, respectively.}

\item{globally}{boolean, default FALSE. should the themes be installed only for the current user (FALSE, default) or for all users (TRUE)}
}
\value{
none
}
\description{
Adds a custom editor theme to RStudio, leveraging the package's internal
dataset with custom RStudio themes. The function also allows to bulk-install
themes.
}
\details{
The function wraps \code{\link{rstudioapi::addTheme}}. It does \emph{not} force
the installation, meaning that existing themes are not overwritten. Also, it
does \emph{not} immediately apply the newly installed theme (see
\code{\link{rstudioapi::applyTheme}}) for this). To see the newly installed
themes, navigate to "Tools"- "Global Options" - "Appearance".
}
\examples{
\dontrun{
# install all dark themes
install_rstudio_themes(theme = "all_dark")

# install two specific themes from the list
install_rstudio_themes(theme = c("Ayu Light", "GitHub {rsthemes}"))
}
}
\seealso{
\code{\link[rstudioapi:addTheme]{rstudioapi::addTheme()}} which this function wraps.
}
